Moved Astyle to src/classes
